Move PersistenceControl to its own file / submodule
